Delete some duplicate code between spares and strikes
scorecard/error fn already checks for the correct number of rolls in the
last frame, so an additional if in scorecard/score-spare and
scorecard/score-strike was not necessary.

In addition, handling spares and strikes in the last frame is the same
(i.e. sum of rolls) so does not need to be handled separately.